requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bleach==6.0.0
Django==4.1.7
django-bootstrap5==22.2
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-summernote==0.8.20.0
Faker==17.6.0
itsdangerous==2.1.2
Pillow==9.4.0
python-dateutil==2.8.2
python-decouple==3.8
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
webencodings==0.5.1
whitenoise==6.4.0
